+++ python/acl_anthology/collections/__init__.py
@@ -18,7 +18,7 @@ from .bibkeys import BibkeyIndex
 from .eventindex import EventIndex
 from .event import Event, Talk
 from .volume import Volume
-from .types import VolumeType
+from .types import EventLinkingType, PaperDeletionType, PaperType, VolumeType
 from .paper import Paper
 
 
@@ -28,7 +28,10 @@ __all__ = [
     "CollectionIndex",
     "Event",
     "EventIndex",
+    "EventLinkingType",
     "Paper",
+    "PaperDeletionType",
+    "PaperType",
     "Talk",
     "Volume",
     "VolumeType",