+++ python/tests/collections/collection_test.py
@@ -17,7 +17,12 @@ from lxml import etree
 from pathlib import Path
 
 from acl_anthology import Anthology
-from acl_anthology.collections import Collection, CollectionIndex, VolumeType
+from acl_anthology.collections import (
+    Collection,
+    CollectionIndex,
+    EventLinkingType,
+    VolumeType,
+)
 from acl_anthology.people import NameSpecification
 from acl_anthology.utils import xml
 from acl_anthology.text import MarkupText
@@ -231,7 +236,9 @@ def test_collection_create_volume_should_create_event(anthology, pre_load, reset
 
     # New implicit event should exist in the event index
     assert "acl-2000" in anthology.events
-    assert volume.full_id_tuple in anthology.events["acl-2000"].colocated_ids
+    assert (volume.full_id_tuple, EventLinkingType.INFERRED) in anthology.events[
+        "acl-2000"
+    ].colocated_ids
     assert volume.full_id_tuple in anthology.events.reverse
     assert anthology.events.reverse[volume.full_id_tuple] == {"acl-2000"}
 
@@ -262,7 +269,9 @@ def test_collection_create_volume_should_update_event(anthology, pre_load, reset
 
     # New volume should be added to existing event
     assert "acl-2022" in anthology.events
-    assert volume.full_id_tuple in anthology.events["acl-2022"].colocated_ids
+    assert (volume.full_id_tuple, EventLinkingType.INFERRED) in anthology.events[
+        "acl-2022"
+    ].colocated_ids
     assert volume.full_id_tuple in anthology.events.reverse
     assert anthology.events.reverse[volume.full_id_tuple] == {"acl-2022"}
 
@@ -331,7 +340,9 @@ def test_collection_create_event_should_update_eventindex(pre_load, anthology):
 
     if pre_load:
         # Volume should automatically have been added
-        assert event.colocated_ids == [collection.get("1").full_id_tuple]
+        assert event.colocated_ids == [
+            (collection.get("1").full_id_tuple, EventLinkingType.INFERRED)
+        ]
     else:
         # If event index wasn't loaded, it's not
         assert event.colocated_ids == []